bin/summary: Use ostree_repo_regenerate_metadata
Call `ostree_repo_regenerate_metadata` with the provided GPG keys to
generate and sign the summary in one call. This changes the handling
when GPGME support is disabled but GPG keys are supplied. Instead of
silently ignoring the option, it will now error. IMO that's better as
callers would otherwise not know that the summary is not GPG signed.