commit: Allow repo to skip automatic summary generation
The ostree-metadata commit is generated within a transaction. If the
repo has auto-update-summary set, then this will trigger the summary to
be generated. However, the only context for creating the ostree-metadata
commit is during summary generation, so this is extra unwanted busy
work. Add a boolean in `OstreeRepoTxn` to disable automatic summary
generation as seen fit.

Currently the commit is created with `ostree summary -u`, which doesn't
have access to the transaction. However, a subsequent commit will move
this into the `ostree_repo_regenerate_metadata` API where it can be
disabled.