lib/commit: Add ostree_repo_get_commit_sizes function
If a commit has the `ostree.sizes` metadata, then an ostree program can
use that data to determine which objects are needed for downloading. The
`ostree_repo_get_commit_sizes` function provides an API for that.

We've been carrying this downstream at Endless for several years and use
it to show accurate download information to users when static deltas
aren't available.